src/requirements.txt
PyMuPDF==1.18.14
matplotlib==3.4.2
numpy==1.19.5
opencv-python==4.4.0.46
path==15.0.0
tensorflow==2.4.0
pyspellchecker==0.6.2
scikit-image==0.18.1
Pillow==8.2.0
flask-restx==0.4.0
pylev==1.3.0
Flask==1.1.4
Werkzeug==1.0.1
lmdb==1.0.0
pytesseract==0.3.7
gunicorn==20.1.0